написала тесты и исправила замечания

+++ src/features/info-calendar/info-calendar-list/ui/info-calendar-list.tsx
@@ -31,7 +31,9 @@ export const InfoCalendarList = (props: InfoCalendarListProps) => {
   }, [dispatch])
 
   const removeCalendarItem = (id: number) => {
-    dispatch(removeCalendar(id))
+    if (window.confirm('Вы точно хотите удалить эту запись?')) {
+      dispatch(removeCalendar(id))
+    }
   }
 
   const filteredCalendars = getUpcomingCalendarDates(calendars).sort(
